src/main/webapp/app/entities/so-huu-theo/delete/so-huu-theo-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { SoHuuTheoService } from '../service/so-huu-theo.service';

import { SoHuuTheoDeleteDialogComponent } from './so-huu-theo-delete-dialog.component';

describe('SoHuuTheo Management Delete Component', () => {
  let comp: SoHuuTheoDeleteDialogComponent;
  let fixture: ComponentFixture<SoHuuTheoDeleteDialogComponent>;
  let service: SoHuuTheoService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [SoHuuTheoDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(SoHuuTheoDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(SoHuuTheoDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(SoHuuTheoService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
